refactor(issue-237): mover InformacionEmpleado a PublicEvents
Primer traslado del Paso 2. Baja al nivel mas bajo del grafo porque es
payload de los tres niveles: DiaCalculado (publico),
ProgramacionTurnoDiarioSolicitada (privado) y TurnoDiarioAsignado (dominio).
