+++ src/Bitakora.ControlAsistencia.Programacion/SolicitarProgramacionTurnoFunction/SolicitarProgramacionTurno.cs
@@ -1,5 +1,5 @@
-using Bitakora.ControlAsistencia.Contracts.Empleados.ValueObjects;
 using Bitakora.ControlAsistencia.Contracts.Programacion.ValueObjects;
+using Bitakora.ControlAsistencia.PublicEvents.Empleados;
 
 namespace Bitakora.ControlAsistencia.Programacion.SolicitarProgramacionTurnoFunction;
